Adicionado DrawerMenu

+++ app/src/main/java/com/codinginflow/despesas/Activity/MainActivity.java
@@ -1,7 +1,9 @@
 package com.codinginflow.despesas.Activity;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 import androidx.recyclerview.widget.ItemTouchHelper;
@@ -30,10 +32,19 @@ public class MainActivity extends AppCompatActivity {
     public static final int EDIT_DESPESA_REQUEST = 2;
     private DespesaViewModel despesaViewModel;
 
+    private DrawerLayout mDrawerLayout;
+    private ActionBarDrawerToggle mToggle;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        mDrawerLayout = (DrawerLayout) findViewById(R.id.drawer);
+        mToggle = new ActionBarDrawerToggle(this, mDrawerLayout, R.string.open, R.string.close);
+        mDrawerLayout.addDrawerListener(mToggle);
+        mToggle.syncState();
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
         // startActivity(new Intent(getApplicationContext(), LoginActivity.class));
 
@@ -135,6 +146,10 @@ public class MainActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
+        if (mToggle.onOptionsItemSelected(item)) {
+            return true;
+        }
+
         switch (item.getItemId()) {
             /*case R.id.delete_all_despesas:
                 despesaViewModel.deleteAllDespesas();
